+++ desktop/cmp/input/Select.ts
@@ -6,7 +6,7 @@
  */
 import {HoistInputModel, HoistInputProps, useHoistInputModel} from '@xh/hoist/cmp/input';
 import {box, div, fragment, hbox, span} from '@xh/hoist/cmp/layout';
-import {createElement, hoistCmp, HoistProps, LayoutProps, PlainObject, XH} from '@xh/hoist/core';
+import {createElement, hoistCmp, HoistProps, LayoutProps, PlainObject, XH, Awaitable} from '@xh/hoist/core';
 import '@xh/hoist/desktop/register';
 import {Icon} from '@xh/hoist/icon';
 import {
@@ -168,7 +168,7 @@ export interface SelectProps extends
      * confused with `filterFn`, which should be used to filter through local options when
      * not in async mode.
      */
-    queryFn?: (query: string) => Promise<SelectOption[]>;
+    queryFn?: (query: string) => Awaitable<(SelectOption|any)[]>;
 
     /**
      * Escape-hatch props passed directly to react-select. Use with care - not all props
@@ -487,32 +487,22 @@ class SelectInputModel extends HoistInputModel {
     //------------------------
     // Async
     //------------------------
-    doQueryAsync = (query) => {
-        return this.componentProps
-            .queryFn(query)
-            .then(matchOpts => {
-                // Normalize query return.
-                matchOpts = this.normalizeOptions(matchOpts);
+    doQueryAsync = async (query) => {
+        const rawOpts = await this.componentProps.queryFn(query),
+            matchOpts = this.normalizeOptions(rawOpts);
+            
+        // Carry forward and add to any existing internalOpts to allow our value
+        // converters to continue all selected values in multiMode.
+        const matchesByVal = keyBy(matchOpts, 'value'),
+            newOpts = [...matchOpts];
+        this.internalOptions.forEach(currOpt => {
+            const matchOpt = matchesByVal[currOpt.value];
+            if (!matchOpt) newOpts.push(currOpt);  // avoiding dupes
+        });
+        this.internalOptions = newOpts;
 
-                // Carry forward and add to any existing internalOpts to allow our value
-                // converters to continue all selected values in multiMode.
-                const matchesByVal = keyBy(matchOpts, 'value'),
-                    newOpts = [...matchOpts];
-
-                this.internalOptions.forEach(currOpt => {
-                    const matchOpt = matchesByVal[currOpt.value];
-                    if (!matchOpt) newOpts.push(currOpt);  // avoiding dupes
-                });
-
-                this.internalOptions = newOpts;
-
-                // But only return the matching options back to the combo.
-                return matchOpts;
-            })
-            .catch(e => {
-                console.error(e);
-                throw e;
-            });
+        // But only return the matching options back to the combo.
+        return matchOpts;
     };
 
     loadingMessageFn = (params) => {