+++ mobx/index.ts
@@ -11,8 +11,6 @@ import {
     computed,
     configure,
     extendObservable,
-    isObservableProp,
-    makeObservable,
     observable,
     override,
     reaction,
@@ -23,7 +21,6 @@ import {
     when
 } from 'mobx';
 import {observer} from 'mobx-react-lite';
-import {bindable, settable} from './decorators';
 
 configure({enforceActions: 'observed'});
 
@@ -33,20 +30,19 @@ configure({enforceActions: 'observed'});
 export {
     action,
     autorun,
-    bindable,
     comparer,
     computed,
     extendObservable,
-    isObservableProp,
-    makeObservable,
     observable,
     observer,
     override,
     reaction,
     runInAction,
-    settable,
     toJS,
     trace,
     untracked,
     when
 };
+
+export * from './decorators';
+export * from './overrides';