+++ promise/Promise.ts
@@ -6,6 +6,7 @@
  */
 import {Exception, ExceptionHandlerOptions, TaskObserver, TrackOptions, XH} from '@xh/hoist/core';
 import {action} from '@xh/hoist/mobx';
+import {olderThan, SECONDS} from '@xh/hoist/utils/datetime';
 import {castArray, isFunction, isNumber, isString} from 'lodash';
 
 /**
@@ -106,6 +107,36 @@ export type PromiseLinkSpec = TaskObserver | { observer: TaskObserver, message?:
  */
 export function wait<T>(interval:number = 0): Promise<T> {
     return new Promise(resolve => setTimeout(resolve, interval)) as Promise<T>;
+}
+
+/**
+ * Return a promise that will resolve after a condition has been met, polling at the specified
+ * interval.
+ *
+ * @param condition - function that should return true when condition is met
+ * @param interval - milliseconds to wait between checks (default 50). Note that the actual time
+ *      will be subject to the minimum delay for `setTimeout()` in the browser.
+ * @param timeout - milliseconds after which the Promise should be rejected (default 5000).
+ */
+export function waitFor(
+    condition: () => boolean,
+    interval: number = 50,
+    timeout: number = 5 * SECONDS
+): Promise<void> {
+    const startTime = Date.now();
+
+    return new Promise((resolve, reject) => {
+        const resolveOnMet = () => {
+            if (condition()) {
+                resolve();
+            } else if (olderThan(startTime, timeout)) {
+                reject(Exception.timeout({interval: Date.now() - startTime}));
+            } else {
+                setTimeout(resolveOnMet, interval);
+            }
+        };
+        resolveOnMet();
+    });
 }
 
 /**